refactor: imgs를 images로 변경
