+++ src/assets/images/index.ts
